fix(decorators): modify delete and options decorators
delete and options decorators now throw error when using body

+++ projects/rest-angular/src/lib/decorators/methods/delete.decorator.ts
@@ -1,6 +1,11 @@
 import {MethodDecoratorFactory} from '../../factories/method-decorator-factory';
+import {NotAllowedDecoratorError} from '../../errors/not-allowed-decorator-error';
 
 export const DELETE = (path: string) => MethodDecoratorFactory.makeDecorator('DELETE', path, (http, request) => {
+  if (request.body) {
+    throw new NotAllowedDecoratorError('Body', 'DELETE');
+  }
+
   return http.delete(
     request.url
   );
